Add Player Health UI and Logic, Adjust Weapon Damage and Enemy Damage

+++ Assets/Scripts/Enemy/EnemyHP.cs
@@ -3,16 +3,16 @@ using UnityEngine;
 public class EnemyHP : MonoBehaviour
 {
     [SerializeField] GameObject robotExplosionVFX;
-    [SerializeField] float startHP = 5f;
+    [SerializeField] int startHP = 5;
 
-    float currentHP;
+    int currentHP;
 
     void Awake()
     {
         currentHP = startHP;
     }
 
-    public void TakeDamage(float amount)
+    public void TakeDamage(int amount)
     {
         currentHP -= amount;
